+++ re-solve.sh
@@ -14,8 +14,8 @@ for json in tr_config/*.seeds.json; do
   echo "================================="
   echo "$video"
   echo $(file -LIb "$video")
-  #./track_runner/track_runner.py --workers 1 -i $video prepare
   echo "================================="
+  ./track_runner/track_runner.py --workers 1 -i $video prepare
   ./track_runner/track_runner.py --workers 2 -i $video solve --yes --bin 1 2>&1 | tee "$log"
   sleep 20
 done